backend/requirements.txt
fastapi
uvicorn
supabase
google-generativeai
python-docx
python-pptx
pydantic
python-dotenv
